about page, simple routing, added config.js
