+++ package.json
@@ -18,6 +18,7 @@
   "homepage": "https://github.com/software-engineering-team-fun/blog-project#readme",
   "dependencies": {
     "bootstrap": "^5.2.2",
+    "express": "^4.18.2",
     "firebase": "^9.12.1",
     "font-awesome": "^4.7.0"
   }
