+++ public/about.html
@@ -1,0 +1,89 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="css/aboutstyle.css">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
+    <title>About</title>
+</head>
+<body>
+  <nav class="navbar navbar-expand-lg bg-dark navbar-dark">
+    <!--Links-->
+    <div class="navbar-nav">
+        <a href="/home" class="nav-item nav-link">Home</a>
+        <a href="" class="nav-item nav-link">About</a>
+        <a href="/login" class="nav-item nav-link">Login</a>
+    </div>
+  </nav>
+    <div class="about-section">
+        <h1>About Us</h1>
+        <p>LiveBlog allows anyone to post a blog and view  other users blog.</p>
+        <p>Utilzing Node.js and Express for server, and Firebase for a database and Hosting</p>
+      </div>
+      
+      <h2 style="text-align:center">Our Team</h2>
+      
+    <div class="row">
+
+        <div class="col-sm">
+          <div class="card">
+            <div class="container">
+              <h2>Maximus Lewis</h2>
+              <p class="title"></p>
+              <p>maximuslewis@lewisu.edu</p>
+              <a href="mailto:some@email.com" target="_blank" class="btn btn-primary">Contact</a>
+            </div>
+          </div>
+        </div>
+
+        <div class="col-sm">
+          <div class="card">
+            <div class="container">
+              <h2>Ryan Johnson</h2>
+              <p class="title"></p>
+              <p>Ryanajohnson1@lewisu.edu</p>
+              <a href="mailto:some@email.com" target="_blank" class="btn btn-primary">Contact</a>
+            </div>
+          </div>
+        </div>
+
+      <div class="col-sm">
+        <div class="card">
+          <div class="container">
+            <h2>Matthew Espinos </h2>
+            <p class="title"></p>
+            <p>matthewwespinos@lewisu.edu</p>
+            <a href="mailto:some@email.com" target="_blank" class="btn btn-primary">Contact</a>
+        </div>
+        </div>
+      </div>
+    </div>
+    <div class="row" id = "bot">
+
+      <div class="col-sm" >
+        <div class="card">
+          <div class="container">
+            <h2>Ryan Syed</h2>
+            <p class="title"></p>
+            <p>ryansyed@lewisu.edu</p>
+            <a href="mailto:some@email.com" target="_blank" class="btn btn-primary">Contact</a>
+        </div>
+        </div>
+      </div>
+        <div class="col-sm">
+          <div class="card">
+            <div class="container">
+              <h2>Anthony Mastores</h2>
+              <p class="title"></p>
+              <p>anthonyjmastores@lewisu.edu</p>
+              <a href="mailto:some@email.com" target="_blank" class="btn btn-primary">Contact</a>
+            </div>
+          </div>
+        </div>
+</div>
+
+</body>
+</html>
+
