+++ public/css/aboutstyle.css
@@ -1,0 +1,46 @@
+body {
+    margin: 0;
+  }
+
+  .col-sm {
+    margin-bottom: 16px;
+    padding: 0 8px; 
+    display: inline-block;    
+
+  }
+  #bot {
+    text-align: center;
+
+  }
+  
+  .card {
+    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
+    margin: 8px;
+  }
+  
+  .about-section {
+    padding: 75px;
+    text-align: center;
+    background-color: #474e5d;
+    color: white;
+  }
+  
+  .container {
+    padding: 0 16px;
+  }
+  
+  .title {
+    color: grey;
+  }
+  
+  .btn-primary {
+    border: none;
+    outline: 0;
+    display: inline-block;
+    padding: 8px;
+    color: white;
+    background-color: #000;
+    text-align: center;
+    cursor: pointer;
+    width: 100%;
+  }
